ai-assist/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-genai
crawl4ai
fastapi-utils
